requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.32.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.4
pandas>=2.2.0
plotly>=5.19.0
